flask/requirements.txt
Flask==2.3.2
requests==2.26.0
python-dotenv==0.19.1
pusher==3.0.0
